pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["Fernando Bonde Leonelli <fernandobondeleonelli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
